dev-requirements.txt
github3.py==0.9.3
httplib2==0.9
launchpadlib==1.6.0
lazr.authentication==0.1.3
lazr.restfulclient==0.13.1
lazr.uri==1.0.3
oauth==1.0.1
requests==2.5.0
simplejson==3.6.5
uritemplate.py==0.3.0
wadllib==1.3.2
wsgi-intercept==0.9.0
wsgiref==0.1.2
zope.interface==4.1.1
